+++ client/changelog.md
@@ -1,3 +1,4 @@
 ### Changes in 2.1.4
 - Fixed `java.lang.NoSuchMethodError: net.minecraft.util.ResourceLocation.func_217855_b(Ljava/lang/String;)Z` on 1.16.5 forge.
 - Fixed issue where voice distance visualization was not fading away with high fps.
+- Fixed entity source icon not being rendered on players.